requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy>=1.21.0
torch>=1.9.0
torchaudio>=0.9.0
transformers==4.31.0
python-dotenv==1.0.0
soundfile>=0.10.3
librosa==0.10.1
pretty_midi>=0.2.10
fluidsynth==0.2
scipy>=1.7.0
pandas==2.0.3
scikit-learn==1.3.0
tqdm==4.65.0 